+++ CHANGELOG.rst
@@ -12,6 +12,26 @@ Changes for the upcoming release can be found in the `"changelog" directory <htt
 
 .. towncrier release notes start
 
+0.4.0 (2023-01-23)
+------------------
+
+
+
+
+Changes
+^^^^^^^
+
+- Updated the signature and docstrings of ``mlnext.plot.plot_signals``.
+  `#32 <https://gitlab.phoenixcontact.com/vmm-factory-automation/digital-factory/data-collection-storage-evaluation/anomaly-detection/mlnext_framework/-/issues/32>`__
+- Refactored ``mlnext.score.eval_sigmoid``.
+  `#33 <https://gitlab.phoenixcontact.com/vmm-factory-automation/digital-factory/data-collection-storage-evaluation/anomaly-detection/mlnext_framework/-/issues/33>`__
+- Removed ``roc_auc`` from ``mlnext.score.eval_metrics``.
+  `#35 <https://gitlab.phoenixcontact.com/vmm-factory-automation/digital-factory/data-collection-storage-evaluation/anomaly-detection/mlnext_framework/-/issues/35>`__
+
+
+----
+
+
 0.3.1 (2022-02-28)
 ------------------